+++ internal/globals.sh
@@ -33,6 +33,9 @@ fi
 [ ! -x "${M4}" ] && echo "m4 not found" && exit 1
 
 [[ "${MULTI_MATCH_STRAT}" == "" ]] && MULTI_MATCH_STRAT="simple"
+[[ "${DEFAULT_INDEX_BY}" == "" ]] && DEFAULT_INDEX_BY="none"
+[[ "${POST_INDEX_BY}" == "" ]] && POST_INDEX_BY="${DEFAULT_INDEX_BY}"
+[[ "${TAG_INDEX_BY}" == "" ]] && TAG_INDEX_BY="${DEFAULT_INDEX_BY}"
 
 function strip_comments {
 	FILE="$1"